+++ VpnHood.Client.App/AppOptions.cs
@@ -29,5 +29,7 @@ public class AppOptions
     public bool LogVerbose { get; set; }
     public bool? LogAnonymous { get; set; }
     public TimeSpan ShowAdPostDelay { get; set; } = TimeSpan.FromSeconds(3);
+    public TimeSpan ServerQueryTimeout { get; set; } = ClientOptions.Default.ServerQueryTimeout;
     public bool SingleLineConsoleLog { get; set; } = true;
+    public bool AutoDiagnose { get; set; } = true;
 }